fix: 修复误删输出图片

+++ electron/src/modules/image-tool/index.ts
@@ -360,6 +360,8 @@ export class ImageTool extends Event {
 
   private delCacheFile() {
     for (const k in this.outputFileNames) {
+      if (k === 'composite') continue;
+
       const _path = (this.outputFileNames as any)[k];
       if (fs.existsSync(_path)) {
         tryCatch(() => fs.rmSync(_path));
